Suppress a bunch of compile warnings.
I couldn't even figure out why my build was failing, because the
actual problem got lost in all the warnings.  This fixes a bunch of
them.

About half of the benefit here is from the one-line change in the
makefile that suppresses a ton of meaningless #pragma warnings.

Most of the rest is from a collection of declared-but-not-used
warnings (mostly variables, but also several functions).  It is
possible that a couple of the unused functions would have been used if
I"d been building for a different platform or if I'd been building the
tcod version (in which case they should be wrapped in #ifdef), but I
wasn't sure, so I commented them out for now.

There were also a couple of places where NULL was passed for a
non-pointer variable, which I'm pretty sure would not actually cause
any problems, because NULL becomes zero when cast to integer, even
implicitly, on every reasonable architecture and platform, I think.
But there's also no legit semantic reason for those function arguments
to be passed as NULL, when what they really are, semantically, is
zero.  They could perhaps be a NO_FLAG macro for zero, but that
really doesn't seem necessary either.  I just made them zero.