Don't specify obsolete 32-bit architecture.
It's 2016.  We've pretty much all been using 64-bit systems for for
about ten years now, and what 32-bit systems do remain in use aren't
x86 (they're mostly arm).

If somebody wants to build for the i386 archictecture for some kind of
retrocomputing project, they'll likely be building _on_ that
architecture, and the compiler will just naturally use it by default.
We certainly don't want it baked into the build system as distributed.

This is the reason my build was failing.  It now succeeds.