Initial basic implementation of roles.
This implements three distinct player roles:
 * adventurer - pot-luck character, item probabilities as in vanilla Brogue
 * barbarian - melee-oriented character, more weapons, no staves
 * mage - more staves and wands, fewer weapons and armor

I have Ideas for several additional roles, but I wanted to get the
basic infrastructure in place first.

The "Welcome" greeting when the game starts mentions your role, so you
always know what role you are.  Any given dungeon seed is always the
same role (and the role can be trivially calculated from the seed
number; in fact, it's not stored separately anywhere, just calculated
from the seed as needed).  I have NOT yet implemented the ability for
the player to choose the role (without choosing a specific seed) when
starting a new game; I want to do that, but I haven't yet.

In addition to the item generation probabilities being different, each
role has a different guarantee:  a particular type of item that will
definitely be generated, eventually (provided you survive long enough
to reach it).  The game tracks whether one of the item in question has
generated yet and, as long as it hasn't, gradually ramps up a special
probability that one of this guaranteed type of item will generate
specially, even if it would not have done so otherwise.  This can only
happen once per game and does not happen at all if the item generates
naturally before it kicks in.  The probability of this happening
reaches an absolute guarantee at dungeon depth twelve, if none of the
item in question have generated yet by then.  There is NOT any special
guarantee about the quality (enchantment) of the item in question; it
can even be cursed and negatively enchanted, if the random numbers
work out that way, following the usual probabilities.  (This might
change in the future; I haven't decided for sure.)

The guarantee is as follows:
 * adventurer - at least one potion of magic detection by depth 12
 * barbarian - at least one ring of transference by depth 12
 * mage - at least one ring of wisdom by depth 12

Currently, role has no impact on vaults or vault items that may be
generated, only random items (such as on the floor).  This may change
in a future commit.

(cherry picked from commit db20722c51493787a3f2307f08c5b148169e919c)