Roles can adjust probabilities of items within a category; add the rogue role
Roles can now adjust the relative probabilities of individual
items within a category.

I wanted to do this anyway (mainly because certain rings are not
useful for barbarians, and other ones are not useful for mages),
but for the rogue role it was absolutely essentail to bias the
armor heavily in favor of lightweight ones.

This commit also implements the rogue role.

(cherry picked from commit 12bcc6d005d7d55aa565924195f3fcf3c368f0fe)