Initial implementation of the healer role.
The healer starts with a staff of healing, of course, and a wand of
domination (because it was easier to implement than guaranteeing a
couple of shackled creatures in the first couple of levels) and no
other weapon (well, except fists and your wits).

If you get early empowerment, which is fairly likely, your chances are
halfway decent.  Until your multi-empowered ally gets itself killed,
of course.

Good luck.